move get code to get mod
